Rename piper to evaluate
