+++ lib/syntax/evaluate.d.ts
@@ -1,0 +1,2 @@
+import { Context, Frame } from "../frames";
+export declare const evaluate: (input: string, context?: Context) => Frame;
