--- lib/syntax/piper.d.ts
@@ -1,2 +1,0 @@
-import { Context, Frame } from "../frames";
-export declare const piper: (input: string, context?: Context) => Frame;
