+++ src/syntax/evaluate.ts
@@ -3,7 +3,7 @@ import { EvalPipe } from "./eval-pipe";
 import { LexPipe } from "./lex-pipe";
 import { ParsePipe } from "./parse";
 
-export const piper = (input: string, context = Void): Frame => {
+export const evaluate = (input: string, context = Void): Frame => {
   const result = new FrameArray([], context); // store the result
   const evaluator = new EvalPipe(result); // evaluate expressions in context
   const parser = new ParsePipe(evaluator); // assemble tokens into expressions
